requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
SQLModel==0.0.16
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
typer[all]==0.12.3
PyYAML==6.0.1
psutil==5.9.8
tomli==2.0.1 ; python_version<"3.11" 